VimHelper uses curses for UI.

+++ VimHelper.py
@@ -5,6 +5,7 @@ import subprocess
 import sys
 import datetime
 import socket
+import curses
 from threading import Thread
 from threading import Condition
 
@@ -15,35 +16,45 @@ class Drawer(Thread):
         self.draw = True
         self.daemon = True
         self.workers = workers
-        self.windowLen = 80
+
+        self.screen = curses.initscr()
+        self.uiOffsetX = 0
+        self.consoleOffsetY = 10
+
+        self.windowWidth = self.screen.getmaxyx()[1]
+        self.windowHeight = self.screen.getmaxyx()[0]
+        self.windowLen = self.windowWidth
         self.msgs = []
-        self.maxMsgs = 20
-        self.consoleNamesLen = 14
-        self.maxTextLen = self.windowLen - self.consoleNamesLen - 6
+        self.maxMsgs = self.windowHeight - self.consoleOffsetY - 2
+        self.consoleNamesLen = 16
+        self.maxTextLen = self.windowLen - self.consoleNamesLen - 4
+        self.barLen = self.windowLen/2 - 2
+        self.maxWorkers = 10
+        curses.noecho()
+        curses.cbreak()
+        self.screen.keypad(1)
+
         for i in range(0,self.maxMsgs):
             self.add_msg("", " ")
-        for worker in self.workers:
-            worker.drawer = self
         self.add_msg("main", "Started vim-helper.")
-        self.barLen = 50
         self.start()
 
     def run(self):
         while(self.draw):
-            os.system('cls')
+            for worker in self.workers:
+                if not worker.isAlive():
+                    self.workers.remove(worker)
             self.drawProgressBars()
             line = '-' * self.windowLen
-            sys.stdout.write("\n")
-            sys.stdout.write(line + "\n")
+            self.screen.addstr(self.consoleOffsetY-1, 0, line)
             self.drawConsole()
-            sys.stdout.write(line + "\n")
-            sys.stdout.flush()
-            time.sleep(0.2)
+            self.screen.addstr(self.windowHeight-2, 0, line)
+            self.screen.refresh()
 
     def add_msg(self, source, msg):
         msgList = msg.split("\n")
         for splitMsg in msgList:
-            source = datetime.datetime.now().strftime("%H:%M:%S|") + source
+            source = datetime.datetime.now().strftime(" %H:%M:%S | ") + source
             tmpMsg = splitMsg
             while len(tmpMsg) > self.maxTextLen:
                 self.msgs.append((source,tmpMsg[0:self.maxTextLen]))
@@ -53,13 +64,52 @@ class Drawer(Thread):
         self.msgs = self.msgs[-self.maxMsgs:]
 
     def drawProgressBars(self):
+        y = 0
+        x = 0
         for worker in self.workers:
-            worker.drawProgressBar()
+            if not worker.hide and not worker.idle and y < 10:
+                count = time.time() - worker.timeStart
+                sleepcount = time.time() - worker.timeDone
+                if worker.idle or worker.lastTime == 0:
+                    status = "Idle..."
+                    percents = 0
+                    filled_len = 0
+                else:
+                    status = "Running"
+                    percents = round(100.0 * count / float(worker.lastTime), 1)
+                    if percents > 99:
+                        percents = 99
+                        filled_len = self.barLen - 1
+                    else:
+                        filled_len = int(round(self.barLen * count / float(worker.lastTime)))
+                bar = ("[" + '#' * filled_len + '-' * (self.barLen - filled_len) + "]")
+
+                self.screen.addstr(y, x*self.barLen, ('{:>10} {} {:>5}'.format("[" + worker.name + "]", status, str(percents) + '%')).center(self.barLen, " "))
+                self.screen.addstr(y + 1, x*(self.barLen + 2), bar)
+                if x == 0:
+                    x += 1
+                else:
+                    x = 0
+                    y += 2
+
+        while y < 10:
+            self.screen.addstr(y, x*self.barLen, "".center(self.barLen, " "))
+            self.screen.addstr(y + 1, x*(self.barLen + 2), "".center(self.barLen + 2, " "))
+            if x == 0:
+                x += 1
+            else:
+                x = 0
+                y += 2
 
     def drawConsole(self):
+        for worker in self.workers:
+            while len(worker.msgs) > 0:
+                self.add_msg(worker.name, worker.msgs.pop(0))
+        i = 0
         for source,msg in self.msgs:
-            consoleLine = "%-" + str(self.consoleNamesLen) + "." + str(self.consoleNamesLen) + "s| %s\n"
-            sys.stdout.write(consoleLine % (source,msg))
+            consoleLine = "%-" + str(self.consoleNamesLen) + "." + str(self.consoleNamesLen) + "s | %s\n"
+            self.screen.addstr(i + self.consoleOffsetY, 0, consoleLine % (source,msg))
+            i += 1
 
 # All external workers should inherit this.
 class Worker(Thread):
@@ -81,12 +131,13 @@ class Worker(Thread):
         self.hide = False
 
         self.condition = Condition()
-        self.sleep = 1
 
-        self.drawer = None
+        self.msgs = []
+
+        self.running = True
 
     def run(self):
-        while(True):
+        while(self.running):
             self.done = False
             self.timeStart = time.time()
             self.update()
@@ -94,8 +145,6 @@ class Worker(Thread):
                 self.timeDone = time.time()
                 self.lastTime = self.timeDone - self.timeStart
                 self.done = True
-            if self.sleep != 0:
-                time.sleep(self.sleep)
 
     def update(self):
         print "fix updateloop"
@@ -104,44 +153,8 @@ class Worker(Thread):
         self.currentPath = path
         self.currentFolder,self.currentFile = os.path.split(path)
 
-    def drawProgressBar(self):
-        if not self.hide:
-            count = time.time() - self.timeStart
-            sleepcount = time.time() - self.timeDone
-            if self.idle or self.lastTime == 0:
-                status = "Idle..."
-                percents = 0
-                filled_len = 0
-            elif self.done:
-                status = "Sleeping..."
-                percents = round(100.0 * sleepcount / float(self.sleep), 1)
-                if percents > 99:
-                    percents = 99
-                    filled_len = 0
-                elif percents < 0:
-                    percents = 0
-                    filled_len = drawer.barLen - 1
-                else:
-                    filled_len = drawer.barLen - int(round(drawer.barLen * sleepcount / float(self.sleep)))
-            else:
-                status = "Running..."
-                percents = round(100.0 * count / float(self.lastTime), 1)
-                if percents > 99:
-                    percents = 99
-                    filled_len = drawer.barLen - 1
-                elif percents < 0:
-                    percents = 0
-                    filled_len = 0
-                else:
-                    filled_len = int(round(drawer.barLen * count / float(self.lastTime)))
-            bar = '#' * filled_len + '-' * (drawer.barLen - filled_len)
-            percentText = str(percents) + '%'
-
-            sys.stdout.write(self.name + '\n')
-            sys.stdout.write('[%s] %s - %s\n\n' % (bar, percentText, status))
-
     def add_msg(self, msg):
-        drawer.add_msg(self.name, msg)
+        self.msgs.append(msg)
 
 # A thread that updates my git statusline
 class UpdateGit(Worker):
@@ -149,7 +162,6 @@ class UpdateGit(Worker):
         Worker.__init__(self, "Git")
         self.daemon = True
         self.lastTime = 0.1
-        self.sleep = 0
         self.pause = True
         self.hide = True
         self.start()
@@ -229,7 +241,6 @@ class TexBuilder(Worker):
         Worker.__init__(self, "LaTex")
         self.daemon = True
         self.lastTime = 3
-        self.sleep = 0
         self.idle = True
         self.start()
 
@@ -238,15 +249,28 @@ class TexBuilder(Worker):
             while self.currentPath == "":
                 self.idle = True
                 self.condition.wait()
-            self.idle = False
+                self.idle = False
             tmpFolder = os.path.expanduser('~') + "/.vim/tmp/tmp/"
             try:
                 if os.path.isfile(tmpFolder + "temp.tex"):
-                    subprocess.check_output("cp " + tmpFolder + "temp.tex " +  tmpFolder + "main.tex", stderr=subprocess.STDOUT)
-                    subprocess.check_output("rm -f " + tmpFolder + "*.aux", stderr=subprocess.STDOUT)
-                    subprocess.check_output("pdflatex -halt-on-error -output-directory=" + tmpFolder + " " + tmpFolder + "main.tex", stderr=subprocess.STDOUT, cwd=self.currentFolder)
+                    if os.path.isfile(tmpFolder + "main.tex"):
+                        f = open(tmpFolder + "temp.tex", 'r')
+                        before = f.read()
+                        f.close()
+                        f = open(tmpFolder + "main.tex", 'r')
+                        after = f.read()
+                        f.close()
+                    if before != after:
+                        self.idle = False
+                        subprocess.check_output("cp " + tmpFolder + "temp.tex " +  tmpFolder + "main.tex", stderr=subprocess.STDOUT)
+                        subprocess.check_output("rm -f " + tmpFolder + "*.aux", stderr=subprocess.STDOUT)
+                        subprocess.check_output("pdflatex -halt-on-error -output-directory=" + tmpFolder + " " + tmpFolder + "main.tex", stderr=subprocess.STDOUT, cwd=self.currentFolder)
+                        self.add_msg("PDF Done")
+                    else:
+                        self.idle = True
             except Exception,e:
-                self.drawer.add_msg(self.name, str(e.output))
+                self.add_msg(str(e.output))
+
 
 class Server(Thread):
     def __init__(self, drawer, workers):
@@ -282,15 +306,17 @@ class Server(Thread):
                 elif server_msgs[1] == "0":
                     self.clients = -1
             elif server_msgs[0] == "tex":
-                if server_msgs[1] == "":
-                    self.add_msg("Stopping PDF")
                 for worker in self.workers:
                     if worker.name == "LaTex":
+                        if server_msgs[1] == "":
+                            worker.add_msg("Stopped listening for changes - " + worker.currentFile)
                         with worker.condition:
                             worker.setPath(server_msgs[1])
                             if server_msgs[1] != "":
-                                self.add_msg("Building PDF - " + worker.currentFile)
+                                worker.add_msg("Listening for changes - " + worker.currentFile)
                             worker.condition.notify()
+            elif server_msgs[0] == "run":
+                self.workers.append(TempBar())
             c.close()
 
     def add_msg(self, msg):
